feat: Add getDateMinusDays utility function
This commit adds the `getDateMinusDays` utility function to the `date.js` module. This function calculates a new date by subtracting a specified number of days from a given date. It is used in the `RecentExpenses` screen to filter expenses that occurred within the last 7 days. This improves the accuracy of the displayed recent expenses.