feat: Update package-lock.json and package.json with axios dependency
This commit adds the axios dependency to the package-lock.json and package.json files. Axios is a popular HTTP client library that will be used for making HTTP requests in the project. This update ensures that the project has the necessary dependencies to use axios and improves the functionality of the project by enabling HTTP communication.